refactor : mvc 패턴으로 패키지 분리
