+++ src/main/java/racingcar/model/Car.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.model;
+
+import racingcar.util.RandomUtil;
 
 public class Car {
     private static final int STANDARD = 4;
