+++ src/main/java/racingcar/util/RandomUtil.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 public interface RandomUtil {
     public int generate();
