+++ src/test/java/racingcar/controller/InputControllerTest.java
@@ -1,6 +1,7 @@
-package racingcar;
+package racingcar.controller;
 
 import org.junit.jupiter.api.Test;
+import racingcar.controller.InputController;
 
 import static org.assertj.core.api.Assertions.assertThat;
