+++ src/test/java/racingcar/controller/RaceControllerTest.java
@@ -1,7 +1,9 @@
-package racingcar;
+package racingcar.controller;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.controller.RaceController;
+import racingcar.model.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
