+++ src/test/java/racingcar/model/CarTest.java
@@ -1,6 +1,7 @@
-package racingcar;
+package racingcar.model;
 
 import org.junit.jupiter.api.Test;
+import racingcar.model.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
