+++ src/test/java/racingcar/util/InputValidatorTest.java
@@ -1,6 +1,7 @@
-package racingcar;
+package racingcar.util;
 
 import org.junit.jupiter.api.Test;
+import racingcar.util.InputValidator;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
